commit zopfli-compressed PNGs for reproducible builds
PNG crunching is not a deterministic process, especially the way aapt does
it.  This makes the builds not reproducible.  The easy solution to this is
to pre-crunch the PNGs and commit them to git.  It also makes the
final APK a tiny amount smaller, since zopfli is an optimized PNG compressor.

https://medium.com/@duhroach/smaller-pngs-and-android-s-aapt-tool-4ce38a24019d